Rename tab.js to be tab-internal.js for clarity.
